% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discordant-package.R
\docType{package}
\name{discordant}
\alias{discordant}
\title{discordant: A Novel Approach for Differential Correlation}
\description{
Discordant is a method to determine differential correlation of molecular 
feature pairs from -omics data using mixture models. Algorithm is explained 
further in Siska et al.
}
\author{
Charlotte Siska

Max McGrath

Katerina Kechris
}
