% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_Breast_miRNASeq}
\alias{TCGA_Breast_miRNASeq}
\title{Example breast miRNA-Seq count dataset.}
\format{
An ExpressionSet with 100 features, 57 samples
}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_Breast_miRNASeq)
}
\description{
This dataset contains TMM normalized voom-transformed miRNA count values from
miRNASeq that was taken from the Cancer Genome Atlas, or TCGA. The dataset 
has 100 miRNA and 57 samples. The original dataset has 212 miRNA and 
57 samples.
}
\examples{
  data(TCGA_Breast_miRNASeq)
}
\references{
National Institutes of Health. The Cancer Genome Atlas.
}
\keyword{datasets}
