% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_Breast_miRNASeq_voom}
\alias{TCGA_Breast_miRNASeq_voom}
\title{Example breast miRNA-Seq voom-transformed count dataset.}
\format{
An ExpressionSet with 100 features and 57 samples
}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_Breast_miRNASeq_voom)
}
\description{
This dataset contains TMM normalized voom-transformed miRNA count values
from miRNASeq that was taken from the Cancer Genome Atlas, or TCGA. The
dataset has 100 miRNA and 57 samples. The original dataset has 212 miRNA and
57 samples.
}
\examples{
  data(TCGA_Breast_miRNASeq_voom)
}
\references{
Charity W Law, Yunshun Chen, Wei Shi, Gordon K Smyth. voom: 
precision weights unlock linear model analysis tools for RNA-seq read 
counts. 2014. Genome Biology, 15:R29.
  
National Institues of Health. The Cancer Genome Atlas.
}
\keyword{datasets}
