\name{TCGA_Breast_RNASeq_voom}
\docType{data}
\alias{TCGA_Breast_RNASeq_voom}
\title{TCGA Breast Cancer RNASeq Sample Dataset}
\description{
	This dataset contains TMM normalized voom-transformed RNA count values from RNASeq that was taken from the Cancer Genome Atlas, or TCGA.
}

\value{
Breast RNA-Seq voom-transformed count data with 100 features and 57 samples.}

\references{
Charity W Law, Yunshun Chen, Wei Shi, Gordon K Smyth. voom: precision weights unlock linear model analysis tools for RNA-seq read counts. 2014. Genome Biology, 15:R29.
National Institues of Health. The Cancer Genome Atlas. http://cancergenome.nih.gov/
}

\author{
Charlotte Siska <siska.charlotte@gmail.com>
}

\usage{TCGA_Breast_RNASeq_voom}
\format{A matrix of RNA count values}
\keyword{datasets}
