% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_GBM_miRNA_microarray}
\alias{TCGA_GBM_miRNA_microarray}
\title{TCGA Glioblastoma Multiforme miRNA Sample Dataset}
\format{
An ExpressionSet with 10 features, 30 samples
}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_GBM_miRNA_microarray)
}
\description{
This dataset contains miRNA expression values from a microarray that was
taken from the Cancer Genome Atlas, or TCGA. It has 10 features and 30
samples. The original dataset had 331 features and 30 samples.
}
\examples{
  data(TCGA_GBM_miRNA_microarray)
}
\references{
National Institutes of Health. The Cancer Genome Atlas.
}
\keyword{datasets}
