\name{TCGA_GBM_transcript_microarray}
\docType{data}
\alias{TCGA_GBM_transcript_microarray}
\title{TCGA Glioblastoma Multiforme Transcript Sample Dataset}
\description{
	This dataset contains transcript expression values from a microarray that was taken from the Cancer Genome Atlas, or TCGA. It has 10 features and 30 samples. The original dataset had 72656 features and 30 samples.
}

\value{
GBM transcript microarray data with 10 features and 30 samples.}

\references{
National Institues of Health. The Cancer Genome Atlas. http://cancergenome.nih.gov/
}

\author{
Charlotte Siska <siska.charlotte@gmail.com>
}

\usage{TCGA_GBM_transcript_microarray}
\format{A matrix of transcript expression values}
\keyword{datasets}
