\name{TCGA_Breast_miRNASeq}
\docType{data}
\alias{TCGA_Breast_miRNASeq}
\title{TCGA Breast Cancer miRNASeq Sample Dataset}
\description{
	This dataset contains TMM normalized miRNA count values from miRNASeq that was taken from the Cancer Genome Atlas, or TCGA. The dataset has 100 miRNA and 57 samples. The original dataset has 212 miRNA and 57 samples.
}

\value{
Breast miRNA-Seq count data with 100 features and 57 samples.}

\references{
National Institues of Health. The Cancer Genome Atlas. http://cancergenome.nih.gov/
}
\author{
Charlotte Siska <siska.charlotte@gmail.com>
}


\usage{TCGA_Breast_miRNASeq}
\format{A matrix of miRNA count values}
\keyword{datasets}
