\name{TCGA_Breast_miRNASeq_voom}
\docType{data}
\alias{TCGA_Breast_miRNASeq_voom}
\title{TCGA Breast Cancer miRNASeq Sample Dataset}
\description{
	This dataset contains TMM normalized voom-transformed miRNA count values from miRNASeq that was taken from the Cancer Genome Atlas, or TCGA. The dataset has 100 miRNA and 57 samples. The original dataset has 212 miRNA and 57 samples.
}

\value{
Breast miRNA-Seq voom-transformed count data with 100 features and 57 samples.}

\references{
Charity W Law, Yunshun Chen, Wei Shi, Gordon K Smyth. voom: precision weights unlock linear model analysis tools for RNA-seq read counts. 2014. Genome Biology, 15:R29.
National Institues of Health. The Cancer Genome Atlas. http://cancergenome.nih.gov/
}

\author{
Charlotte Siska <siska.charlotte@gmail.com>
}

\usage{TCGA_Breast_miRNASeq_voom}
\format{A matrix of miRNA count values}
\keyword{datasets}
