\name{TCGA_Breast_RNASeq}
\docType{data}
\alias{TCGA_Breast_RNASeq}
\title{TCGA Breast Cancer RNASeq Sample Dataset}
\description{
	This dataset contains TMM normalized RNA count values from RNASeq that was taken from the Cancer Genome Atlas, or TCGA. It has 100 features and 57 samples. The original dataset had 17972 features and 57 samples.
}

\references{
National Institues of Health. The Cancer Genome Atlas. http://cancergenome.nih.gov/
}

\author{
Charlotte Siska <siska.charlotte@gmail.edu>
}

\usage{TCGA_Breast_RNASeq}
\format{A matrix of RNA count values}
\keyword{datasets}
