% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_Breast_RNASeq_voom}
\alias{TCGA_Breast_RNASeq_voom}
\title{TCGA Breast Cancer RNASeq Sample Dataset}
\format{
An ExpressionSet with 100 features and 57 samples
}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_Breast_miRNASeq_voom)
}
\description{
This dataset contains TMM normalized voom-transformed RNA count values from
RNASeq that was taken from the Cancer Genome Atlas, or TCGA.
}
\examples{
  data(TCGA_Breast_miRNASeq_voom)
}
\references{
Charity W Law, Yunshun Chen, Wei Shi, Gordon K Smyth. voom: precision 
weights unlock linear model analysis tools for RNA-seq read counts. 2014. 
Genome Biology, 15:R29.

National Institues of Health. The Cancer Genome Atlas.
}
\keyword{datasets}
