\name{TCGA_GBM_miRNA_microarray}
\docType{data}
\alias{TCGA_GBM_miRNA_microarray}
\title{TCGA Glioblastoma Multiforme miRNA Sample Dataset}
\description{
	This dataset contains miRNA expression values from a microarray that was taken from the Cancer Genome Atlas, or TCGA. It has 10 features and 30 samples. The original dataset had 331 features and 30 samples.
}

\value{
GBM miRNA microarray data with 10 features and 30 samples.}

\references{
National Institues of Health. The Cancer Genome Atlas. http://cancergenome.nih.gov/
}

\author{
Charlotte Siska <siska.charlotte@gmail.com>
}

\usage{TCGA_GBM_miRNASample}
\format{A matrix of miRNA expression values}
\keyword{datasets}
